Feature: Adds etag cache validation.
The main idea behind this is how can we avoid downloading again zip files that are already present in the local sub-folder.

Of course, by name might do the trick but downloaded zips could be corrupted, before we did a checksum over the downloaded content and the local zip, while this is valid, it still require us to download the whole zip.

We now get the etag and store it with the name of the repo. We will use a head request to get this etag and if we match the zip we will give it as valid.