Fix race condition when we ran multiple sequential dedup commands.
The filecount variable was outside the match clause (this might be a problem but I suspect it was not as we only handle the case where fc == 0 and it is unlikely to have a read error on the first record).
We flush the writer after each write_repo_jsonl call.